Refactor credential requests
Adds a new hook `useAuth` that requests credentials for subsequent requests made from a component. The intent is to replace the numerous `authApi` calls made on every request with just a single call made when the component mounts. This doesn't replace all instances of the `authApi` calls because some of those happen outside of the component, but this change should still reduce the QPS the plugin is making.