Add integration Tests (#54)
* Add simple integration test to ensure settings menu is available

* Fix project list to dynamically search for results

* add allow-root to integration tests

* Add license headers for new files

* Add MUI and use MUI for config selection

We should move off of semantic UI due to global css pollution concerns.

* Add MUI and use MUI for config selection

We should move off of semantic UI due to global css pollution concerns.

* Refactor authentication in clusters to use new "authenticatedFetch" to wrap up all the auth logic outside of the component. (#44)

Adds a wrapper to fetch to call the authApi and handle setting up authorization info for the requests.

Refactors the use of authApi() in clusters to use the new wrapper. Subsequent changes will add changes to the other components.

Removed a reference to the listClusterApi that was unused.

* Add authenticated fetch to some components (#49)

* Adding new Session Template Form (#52)

This was getting a little large so I'm checking this in-progress
form that isn't currently hooked up yet. (All the Edit*Control files
are not currently being used outside of Storybook), but this PR also
contains a number of additional changes.

- Generalized ProjectIdDropdown so it could be reused by other dynamic
  dropdowns.
- Added dependencies to Storybook, Formik and Yup for workbenching,
  form valdidation and schema-ing respectively.
- Added a couple of services.tsx files that aren't used yet.

* generate sponge_log.xml

---------

Co-authored-by: James Lee <4108254+jinnthehuman@users.noreply.github.com>