Fail gracefully at startup if the plugin prerequisites are not met.
This change modifies the statup hooks for the Dataproc Jupyter Plugin so that
it gracefully degrades if any of the requirements for it are not met.

Specifically, with this change the plugin will verify that the Gateway Client
URL used with Dataproc can be correctly identified before configuring the
Jupyter server to use that URL.

When this cannot happen (either because the project or region are not set
or because the user is not logged in to gcloud with an account that has
sufficient permissions), then we fall back to only supporting local kernels,
although the various UI components of the Dataproc Jupyter Plugin are still
available.

That fallback means that the user can still make full use of their local
kernels even though they cannot use remote kernels.

Additionally, this change extends the config handler so that it reports
an error (via the fields in its returned JSON payload) in case the
setup failed, which will allow us to extend the UI to display a warning
to the user in that scenario.