Fix Issue with comparing config flags that are unset
Unset config flags are set to LazyConfigValue type which does not support int comparison like `max`. The change adds a check to make sure the config value is set before doing a `max` comparison. If the config value is not set, the it just sets the value to the predefined min.

Also fix an issue with gateway_request_timeout not being a real flag name. Renamed it to request_timeout.